requirements.txt
Django>=3.0.4,<3.1.0
djangorestframework>=3.11.0,<3.12.0
psycopg2-binary==2.8.4
flake8==3.7.9
Pillow>=7.1.1,<7.2.0
coverage>=5.0.4,<5.1.0
